Drobne poprawki

+++ src/Maciek/Fsd.Maciek.Cs/Fsd.Maciek.Cs.Ex4.Data/Models/Product.cs
@@ -6,11 +6,11 @@ namespace Fsd.Maciek.Cs.Ex4.Data.Models
     {
         public string Manufacturer { get; private set; }
         public string Model { get; private set; }
-        public float Price { get; private set; }
+        public decimal Price { get; private set; }
         public int ProductionYear { get; private set; }
         public ProductType ProductType { get; private set; }
 
-        public Product(string manufacturer, string model, float price, int productionYear, ProductType productType)
+        public Product(string manufacturer, string model, decimal price, int productionYear, ProductType productType)
         {
             Manufacturer = manufacturer;
             Model = model;
