+++ src/Maciek/Fsd.Maciek.Cs/Fsd.Maciek.Cs.Ex4.Web/Models/ProductWebModel.cs
@@ -4,7 +4,7 @@
     {
         public string Manufacturer { get; internal set; }
         public string Model { get; internal set; }
-        public float Price { get; internal set; }
+        public decimal Price { get; internal set; }
         public int ProductionYear { get; internal set; }
         public string ProductType { get; internal set; }
     }
